requirements.txt
pandas==2.1.2
nltk==3.8.1
Flask-Cors==4.0.0
Flask==3.0.0
gensim==4.3.2
spacy==3.7.4
textblob==0.18.0.post0
scikit-learn==1.3.2
yake==0.4.8
requests==2.31.0
torch==2.2.1
urllib3==2.0.7
kaggle==1.6.11
nbformat==5.10.4
kaggle==1.6.11
google-ai-generativelanguage==0.6.1